Move completion to pkg/cli/completion
