+++ pkg/cli/completion/completion.go
@@ -1,4 +1,4 @@
-package cmd
+package completion
 
 import (
 	"context"
@@ -17,39 +17,16 @@ const completionTimeout = time.Second * 3
 
 // defining as a type purely for readability purposes.
 // this is the type accepted by cobra.Command.ValidArgsFunc and cobra.Command.RegisterFlagCompletionFunc
-type completionFunc func(*cobra.Command, []string, string) ([]string, cobra.ShellCompDirective)
+type Func func(*cobra.Command, []string, string) ([]string, cobra.ShellCompDirective)
 
 type completionResult struct {
 	completions []string
 	directive   cobra.ShellCompDirective
 }
 
-// wrapper to add a timeout to completionFuncs
-func wrapCompletionFuncWithTimeout(defaultDirective cobra.ShellCompDirective, compFunc completionFunc) completionFunc {
-	return func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-		// initialize a buffered channel to receive results
-		resultChan := make(chan completionResult, 1)
-
-		// run completion function in the background and send result to resultChan
-		go func(c chan completionResult) {
-			completions, directive := compFunc(cmd, args, toComplete)
-			r := completionResult{completions: completions, directive: directive}
-			c <- r
-		}(resultChan)
-
-		// wait for results or timeout
-		select {
-		case result := <-resultChan:
-			return result.completions, result.directive
-		case <-time.After(completionTimeout):
-			return []string{}, defaultDirective | cobra.ShellCompDirectiveError
-		}
-	}
-}
-
-// newValidVClusterNameFunc returns a function that handles shell completion when the argument is vcluster_name
+// NewValidVClusterNameFunc returns a function that handles shell completion when the argument is vcluster_name
 // It takes into account the namespace if specified by the --namespace flag.
-func newValidVClusterNameFunc(globalFlags *flags.GlobalFlags) completionFunc {
+func NewValidVClusterNameFunc(globalFlags *flags.GlobalFlags) Func {
 	fn := func(cmd *cobra.Command, _ []string, _ string) ([]string, cobra.ShellCompDirective) {
 		vclusters, err := find.ListVClusters(cmd.Context(), globalFlags.Context, "", globalFlags.Namespace, log.Default.ErrorStreamOnly())
 		if err != nil {
@@ -64,8 +41,8 @@ func newValidVClusterNameFunc(globalFlags *flags.GlobalFlags) completionFunc {
 	return wrapCompletionFuncWithTimeout(cobra.ShellCompDirectiveNoFileComp, fn)
 }
 
-// newNamespaceCompletionFunc handles shell completions for the namespace flag
-func newNamespaceCompletionFunc(ctx context.Context) completionFunc {
+// NewNamespaceCompletionFunc handles shell completions for the namespace flag
+func NewNamespaceCompletionFunc(ctx context.Context) Func {
 	fn := func(_ *cobra.Command, _ []string, _ string) ([]string, cobra.ShellCompDirective) {
 		restConfig, err := config.GetConfig()
 		if err != nil {
@@ -90,4 +67,27 @@ func newNamespaceCompletionFunc(ctx context.Context) completionFunc {
 		return names, cobra.ShellCompDirectiveNoFileComp
 	}
 	return wrapCompletionFuncWithTimeout(cobra.ShellCompDirectiveNoFileComp, fn)
+}
+
+// wrapper to add a timeout to completionFuncs
+func wrapCompletionFuncWithTimeout(defaultDirective cobra.ShellCompDirective, compFunc Func) Func {
+	return func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+		// initialize a buffered channel to receive results
+		resultChan := make(chan completionResult, 1)
+
+		// run completion function in the background and send result to resultChan
+		go func(c chan completionResult) {
+			completions, directive := compFunc(cmd, args, toComplete)
+			r := completionResult{completions: completions, directive: directive}
+			c <- r
+		}(resultChan)
+
+		// wait for results or timeout
+		select {
+		case result := <-resultChan:
+			return result.completions, result.directive
+		case <-time.After(completionTimeout):
+			return []string{}, defaultDirective | cobra.ShellCompDirectiveError
+		}
+	}
 }
